В отдельный модуль utils.TestData вынесены: список групп. В остальных файлах вместо хранения списка групп по отдельности, теперь используется централизованный вызов
