+++ src/routers/create_new_schedule.py
@@ -1,7 +1,7 @@
 import random
 import datetime
 from utils.date import get_day_of_the_week_by_date
-from utils.TestData import list_of_lesson_name, list_of_time_lesson, list_of_type_lesson, list_of_teacher_name
+from utils.TestData import list_of_lesson_name, list_of_time_lesson, list_of_type_lesson, list_of_teacher_name, get_all_group
 class PairTeacherLesson:
     def __init__(self, teacher, lesson):
         self.teacher = teacher
@@ -21,7 +21,7 @@ def generate_pair_teacher_lesson():
     return list_of_pair_lesson_teacher
 
 
-def generate_schedule_in_day(list_of_pair_lesson_teacher, group:str, number_lesson: int, number_of_week: int, day: int):
+def generate_schedule_for_day(list_of_pair_lesson_teacher, group:str, number_lesson: int, number_of_week: int, day: int):
     k = random.randint(0, len(list_of_pair_lesson_teacher)-1)
     date = datetime.datetime.strptime('2022-09-05', "%Y-%m-%d")
     date = date+datetime.timedelta(weeks=number_of_week, days=day)
@@ -41,18 +41,17 @@ def generate_schedule_in_day(list_of_pair_lesson_teacher, group:str, number_less
 
 
 def create_fake_schedule():
-    data = []
-    list_of_group = ["МО-111", "ПРО-123", "МО-211", "ПРО-229", "МО-333", "ПРО-329", "И-111", "М-123", "И-222", "М-223",
-                     "И-333", "М-323", "МО-444", "ПРО-429","И-444", "М-423"]
+    fake_schedule = []
+    list_of_group = get_all_group()
     list_of_pair_teacher_lesson = generate_pair_teacher_lesson()
     for group in list_of_group:
         for week in range(0,10):
             for day in range(0,6):
                 count_of_lesson = random.randint(1,4)
                 for number_of_lesson in range(1, count_of_lesson):
-                    schedule_data = generate_schedule_in_day(list_of_pair_teacher_lesson,group, number_of_lesson, week, day)
-                    data.append(schedule_data)
-    return data
+                    schedule_for_day = generate_schedule_for_day(list_of_pair_teacher_lesson, group, number_of_lesson, week, day)
+                    fake_schedule.append(schedule_for_day)
+    return fake_schedule
 
 if __name__ == '__main__':
     with open('test_data.txt','w') as file:
